add declarations
Declarations are references to outside definitions, which can behave as
you want. Effectively, it's like a foreign function interface for types,
meaning we can refer to an outside definition by declaring where it
exists and relying on it having the correct definitions. This means we
can embed more advanced types but not define special behavior for them,
instead relying on the user to supply the implementation for each
language they're interested in.