[py] Change encoder output
I was trying to be lazy and have the encoding be done mostly behind the
scenes in Python, but this turned out to be a pipe dream. Right now it's
a lot like the F# output in that we output each field. I'm still using
`json.dumps` for the actual output, so we can pass the literal values in
there in most cases, but for nested structures we have to call their
encoders (and we special-case some of the complex types as well).