make parser much less rigid about whitespace
It remains to be seen whether or not most of these are good in practice,
but I think this will mean less frustration for most new users. In any
case users really should keep a common formatting in their project,
though, but at least now they are not forced into a very, very specific
format by the parser.