feat(kt): add companion `create` function for structs
This allows us to create an instance of a struct where all the optional
fields are automatically set to `null`, without having to poison the
original definition with these default values. This means we can use
`create` situationally when it makes sense to create an empty version of
a struct and then fill in fields as we go, but still enforce fields
existing when using the normal constructors (which are otherwise more
convenient and direct).